This is effectively a total rewrite of the IM service, but I've somehow made the Python interface almost identical (with the addition of a couple of conversation-oriented methods).
The new IM service is now based around individual conversations, which have a
set of capabilities. Some conversations support more functions, such as HTML
and subject messages, while others adhere to the older available messages.
This is all based off of how the NotesBuddy client appears to do it.

I've also added a mw_ prefix to all of the installed headers, to ensure that
there's no conflict with two headers of the same name in different packages,
since I used rather common words for some of the header names.