the start of some large changes, with of course more to come. Putting in some foreach functionality for channels and ciphers, since that's pretty much the only thing the GList* version was useful for anyway. The GClosure version of the foreach functions should be useful in the wrappers, as well, if only for aiding in generating the appropriate collection type.
thus far, I'm dropping SWIG for PyGTK and whatever Perl uses that's similar to
PyGTK. SWIG was too much of a pain to adapt to the GObject heirarchy and its
virtual methods and signals. But then PyGTK doesn't provide a simple method
for providing a type conversion (and there's nothing built-in for GList* or
gchar**, how annoying is that?) while SWIG does.

Maybe my next project will be an attempt to provide a new wrapper generator
that does all the things I really want.